# packages/openapi/spec/admin/media.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Voyant Framework API",
    "version": "0.0.0",
    "description": "Generated from the Voyant framework's standard module composition. Do not edit by hand."
  },
  "servers": [
    {
      "url": "/",
      "description": "This deployment (same origin)"
    }
  ],
  "components": {
    "schemas": {
      "CatalogSearchFilter": {
        "anyOf": [
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "eq"
                ]
              },
              "field": {
                "type": "string",
                "minLength": 1
              },
              "value": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  }
                ]
              }
            },
            "required": [
              "kind",
              "field",
              "value"
            ]
          },
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "in"
                ]
              },
              "field": {
                "type": "string",
                "minLength": 1
              },
              "values": {
                "type": "array",
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "number"
                    }
                  ]
                }
              }
            },
            "required": [
              "kind",
              "field",
              "values"
            ]
          },
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "range"
                ]
              },
              "field": {
                "type": "string",
                "minLength": 1
              },
              "gte": {
                "type": "number"
              },
              "lte": {
                "type": "number"
              }
            },
            "required": [
              "kind",
              "field"
            ]
          },
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "and"
                ]
              },
              "clauses": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CatalogSearchFilter"
                }
              }
            },
            "required": [
              "kind",
              "clauses"
            ]
          },
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "or"
                ]
              },
              "clauses": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CatalogSearchFilter"
                }
              }
            },
            "required": [
              "kind",
              "clauses"
            ]
          }
        ]
      },
      "BookingProductDetailsInput": {
        "type": "object",
        "properties": {
          "productId": {
            "type": [
              "string",
              "null"
            ]
          },
          "optionId": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "BookingItemProductDetailsInput": {
        "type": "object",
        "properties": {
          "productId": {
            "type": [
              "string",
              "null"
            ]
          },
          "optionId": {
            "type": [
              "string",
              "null"
            ]
          },
          "unitId": {
            "type": [
              "string",
              "null"
            ]
          },
          "supplierServiceId": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      }
    },
    "parameters": {}
  },
  "paths": {
    "/v1/admin/uploads": {
      "post": {
        "summary": "Upload media",
        "x-voyant-api-id": "@voyant-travel/storage#api.admin.uploads",
        "responses": {
          "200": {
            "description": "The stored media reference."
          },
          "400": {
            "description": "The multipart request is invalid."
          },
          "413": {
            "description": "The upload exceeds the configured limit."
          },
          "415": {
            "description": "The uploaded media type is not supported."
          },
          "503": {
            "description": "Object storage is not configured."
          }
        },
        "operationId": "postAdminUploads",
        "tags": [
          "media"
        ],
        "x-voyant-module": "media",
        "x-voyant-surface": "admin"
      }
    },
    "/v1/admin/uploads/video": {
      "post": {
        "summary": "Create a video upload ticket",
        "x-voyant-api-id": "@voyant-travel/storage#api.admin.video-upload-ticket",
        "responses": {
          "200": {
            "description": "A provider-specific video upload ticket."
          },
          "400": {
            "description": "The ticket request is invalid."
          }
        },
        "operationId": "postAdminUploadsVideo",
        "tags": [
          "media"
        ],
        "x-voyant-module": "media",
        "x-voyant-surface": "admin"
      }
    },
    "/v1/admin/media/{key}": {
      "get": {
        "summary": "Download stored media",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "allowReserved": true,
            "description": "Nested object key under uploads/ or brochures/products/.",
            "schema": {
              "type": "string",
              "pattern": "^(?:uploads|brochures/products)/.+$"
            }
          }
        ],
        "x-voyant-api-id": "@voyant-travel/storage#api.admin.media",
        "responses": {
          "200": {
            "description": "The stored media bytes."
          },
          "400": {
            "description": "The storage key is invalid."
          },
          "404": {
            "description": "The object does not exist."
          },
          "503": {
            "description": "Object storage is not configured."
          }
        },
        "operationId": "getAdminMediaByKey",
        "tags": [
          "media"
        ],
        "x-voyant-module": "media",
        "x-voyant-surface": "admin"
      }
    }
  },
  "webhooks": {}
}